done 1, refactoring cities
