+++ src/const.ts
@@ -1,4 +1,4 @@
-import {cityMap} from './mocks/city.ts';
+import {Location} from './types/location.ts';
 
 export enum AppRoute {
   Main = '/',
@@ -30,16 +30,64 @@ export const URL_MARKER_CURRENT = '../public/img/pin-active.svg';
 
 export const NEARBY_OFFERS_LIMIT = 3;
 
-export const CITY_NAMES = [
-  'Paris',
-  'Cologne',
-  'Brussels',
-  'Amsterdam',
-  'Hamburg',
-  'Dusseldorf',
-] as const;
+export const CITIES: Record<string, { name: string; location: Location }> = {
+  Paris: {
+    name: 'Paris',
+    location: {
+      latitude: 48.85661,
+      longitude: 2.351499,
+      zoom: 13,
+    },
+  },
+  Cologne: {
+    name: 'Cologne',
+    location: {
+      latitude: 50.938361,
+      longitude: 6.959974,
+      zoom: 13,
+    },
+  },
+  Brussels: {
+    name: 'Brussels',
+    location: {
+      latitude: 50.846557,
+      longitude: 4.351697,
+      zoom: 13,
+    },
+  },
+  Amsterdam: {
+    name: 'Amsterdam',
+    location: {
+      latitude: 52.37454,
+      longitude: 4.897976,
+      zoom: 13,
+    },
+  },
+  Hamburg: {
+    name: 'Hamburg',
+    location: {
+      latitude: 53.550341,
+      longitude: 10.000654,
+      zoom: 13,
+    },
+  },
+  Dusseldorf: {
+    name: 'Dusseldorf',
+    location: {
+      latitude: 51.225402,
+      longitude: 6.776314,
+      zoom: 13,
+    },
+  },
+} as const;
 
-export const DEFAULT_CITY = cityMap.Paris;
+export type CityName = keyof typeof CITIES;
+
+export const CITY_NAMES = Object.keys(CITIES) as readonly CityName[];
+
+export const DEFAULT_CITY_NAME: CityName = CITY_NAMES[0];
+
+export const DEFAULT_CITY = CITIES[DEFAULT_CITY_NAME];
 
 export enum SortingOption {
   Popular = 'Popular',
