+++ src/index.tsx
@@ -4,12 +4,13 @@ import App from './app/app.tsx';
 import {Provider} from 'react-redux';
 import {store} from './store';
 import {loadOffers} from './store/action.ts';
-
-import {offers} from './mocks/offers.ts';
+import {Offer} from './types/offer.ts';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
+
+const offers: Offer[] = [];
 
 store.dispatch(loadOffers(offers));
