--- src/mocks/offers.ts
@@ -1,206 +1,0 @@
-import {Offer} from '../types/offer.ts';
-import {cityMap} from './city';
-
-export const offers: Offer[] = [
-  {
-    id: '8f6a4dbd-6579-4da5-b0b4-67bdefda61d8',
-    title: 'Perfectly located Castro',
-    description: 'I am happy to welcome you to my apartment in the city center! Three words: location, cosy and chic!',
-    type: 'apartment',
-    price: 251,
-    previewImage: 'https://14.design.htmlacademy.pro/static/hotel/10.jpg',
-    images: [
-      'https://14.design.htmlacademy.pro/static/hotel/1.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/19.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/11.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/10.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/15.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/17.jpg'
-    ],
-    city: cityMap.Paris,
-    location: {
-      latitude: 48.868610000000004,
-      longitude: 2.342499,
-      zoom: 16
-    },
-    goods: [
-      'Baby seat',
-      'Wi-Fi',
-      'Washer',
-      'Washing machine'
-    ],
-    host: {
-      isPro: true,
-      name: 'Angelina',
-      avatarUrl: 'https://14.design.htmlacademy.pro/static/host/avatar-angelina.jpg'
-    },
-    isPremium: true,
-    isFavorite: true,
-    rating: 2.5,
-    bedrooms: 4,
-    maxAdults: 7
-  },
-  {
-    id: '101cf67d-35bb-4d51-bec2-00658e49e710',
-    title: 'Amazing and Extremely Central Flat',
-    description: 'Design interior in most sympathetic area! Complitely renovated, well-equipped, cosy studio in idyllic, over 100 years old wooden house. Calm street, fast connection to center and airport.',
-    type: 'room',
-    price: 260,
-    previewImage: 'https://14.design.htmlacademy.pro/static/hotel/5.jpg',
-    images: [
-      'https://14.design.htmlacademy.pro/static/hotel/15.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/2.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/10.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/16.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/11.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/7.jpg'
-    ],
-    city: cityMap.Cologne,
-    location: {
-      latitude: 50.950361,
-      longitude: 6.959974,
-      zoom: 16
-    },
-    goods: [
-      'Laptop friendly workspace',
-      'Washer',
-      'Dishwasher',
-      'Wi-Fi',
-      'Kitchen',
-      'Coffee machine',
-      'Breakfast'
-    ],
-    host: {
-      isPro: true,
-      name: 'Angelina',
-      avatarUrl: 'https://14.design.htmlacademy.pro/static/host/avatar-angelina.jpg'
-    },
-    isPremium: true,
-    isFavorite: false,
-    rating: 4.3,
-    bedrooms: 1,
-    maxAdults: 3
-  },
-  {
-    id: '183e3f50-27f9-4d2e-bbd1-c1190c884a3c',
-    title: 'Wood and stone place',
-    description: 'A new spacious villa, one floor. All commodities, jacuzzi and beautiful scenery. Ideal for families or friends.',
-    type: 'house',
-    price: 518,
-    previewImage: 'https://14.design.htmlacademy.pro/static/hotel/10.jpg',
-    images: [
-      'https://14.design.htmlacademy.pro/static/hotel/14.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/4.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/18.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/9.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/2.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/16.jpg'
-    ],
-    city: cityMap.Brussels,
-    location: {
-      latitude: 50.854557,
-      longitude: 4.364697,
-      zoom: 16
-    },
-    goods: [
-      'Baby seat',
-      'Washing machine',
-      'Wi-Fi',
-      'Breakfast'
-    ],
-    host: {
-      isPro: true,
-      name: 'Angelina',
-      avatarUrl: 'https://14.design.htmlacademy.pro/static/host/avatar-angelina.jpg'
-    },
-    isPremium: true,
-    isFavorite: false,
-    rating: 2.4,
-    bedrooms: 4,
-    maxAdults: 2
-  },
-  {
-    id: '98654b55-60c3-4f7f-b03e-10cb3a67a0c7',
-    title: 'The house among olive ',
-    description: 'I am happy to welcome you to my apartment in the city center! Three words: location, cosy and chic!',
-    type: 'apartment',
-    price: 112,
-    previewImage: 'https://14.design.htmlacademy.pro/static/hotel/2.jpg',
-    images: [
-      'https://14.design.htmlacademy.pro/static/hotel/6.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/13.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/7.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/5.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/18.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/3.jpg'
-    ],
-    city: cityMap.Amsterdam,
-    location: {
-      latitude: 52.3909553943508,
-      longitude: 4.939309666406198,
-      zoom: 16
-    },
-    goods: [
-      'Air conditioning',
-      'Wi-Fi',
-      'Coffee machine',
-      'Washing machine',
-      'Towels',
-      'Cable TV'
-    ],
-    host: {
-      isPro: true,
-      name: 'Angelina',
-      avatarUrl: 'https://14.design.htmlacademy.pro/static/host/avatar-angelina.jpg'
-    },
-    isPremium: true,
-    isFavorite: false,
-    rating: 4.4,
-    bedrooms: 2,
-    maxAdults: 9
-  },
-  {
-    id: '038d31f2-f169-4177-b2cd-d9e4ffa52f25',
-    title: 'Penthouse, 4-5 rooms + 5 balconies',
-    description: 'I rent out a very sunny and bright apartment only 7 minutes walking distance to the metro station. The apartment has a spacious living room with a kitchen, one bedroom and a bathroom with mit bath. A terrace can be used in summer.',
-    type: 'room',
-    price: 161,
-    previewImage: 'https://14.design.htmlacademy.pro/static/hotel/1.jpg',
-    images: [
-      'https://14.design.htmlacademy.pro/static/hotel/15.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/18.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/17.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/9.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/6.jpg',
-      'https://14.design.htmlacademy.pro/static/hotel/3.jpg'
-    ],
-    city: cityMap.Paris,
-    location: {
-      latitude: 48.858610000000006,
-      longitude: 2.330499,
-      zoom: 16
-    },
-    goods: [
-      'Dishwasher',
-      'Laptop friendly workspace',
-      'Towels',
-      'Kitchen',
-      'Cable TV',
-      'Wi-Fi',
-      'Air conditioning',
-      'Washing machine',
-      'Heating',
-      'Baby seat'
-    ],
-    host: {
-      isPro: true,
-      name: 'Angelina',
-      avatarUrl: 'https://14.design.htmlacademy.pro/static/host/avatar-angelina.jpg'
-    },
-    isPremium: true,
-    isFavorite: false,
-    rating: 3.8,
-    bedrooms: 1,
-    maxAdults: 3
-  }
-];
