+++ src/types/city.ts
@@ -1,7 +1,5 @@
 import {Location} from './location.ts';
-import {CITY_NAMES} from '../const.ts';
-
-export type CityName = (typeof CITY_NAMES)[number];
+import {CityName} from '../const.ts';
 
 export type City = {
   name: CityName;
